move leveldesign into its own directory. cleanup includes.
