+++ source/logic/world/leveldesign/CuboidChunk.cpp
@@ -1,7 +1,5 @@
 #include "CuboidChunk.h"
 
-#include "Cuboid.h"
-
 CuboidChunk::CuboidChunk()
 {
 }
